wony_test CRUD complete

+++ dev/workspace/.metadata/.plugins/org.eclipse.wst.server.core/tmp0/wtpwebapps/wony_test/view.jsp
@@ -114,8 +114,8 @@
 			<%
 				if(userID != null && userID.equals(bbs.getUserID())){
 			%>
-					<a href="update.jsp??bbsID=<%=bbsID%>" class="btn btn-primary">수정</a>
-					<a href="deleteAction.jsp??bbsID=<%=bbsID%>" class="btn btn-primary">삭제</a>
+					<a href="update.jsp?bbsID=<%=bbsID%>" class="btn btn-primary">수정</a>
+					<a onclick="return confirm('정말로 삭제하겠습니까?')" href="deleteAction.jsp?bbsID=<%=bbsID%>" class="btn btn-primary">삭제</a>
 			<%
 					}
 			%>
